fix: don't let delete logged in user

+++ app/Http/Controllers/Admin/UserController.php
@@ -7,6 +7,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Notifications\NewUser;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Kreait\Firebase\Factory;
 use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Messaging\Notification;
@@ -87,6 +88,10 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
+
+        if (Auth::user()->id === $user->id) {
+            return redirect()->route('users.index')->with('error', 'Você não pode excluir a si mesmo!');
+        }
 
         $user->roles()->detach();
         $user->delete();
